requirements.txt
streamlit
pandas
gspread
gspread-dataframe
google-auth
oauth2client
vnstock
